facade_service/requirements.txt
fastapi
uvicorn[standard]
requests
hazelcast-python-client
pydantic
pydantic-factories